Simon_game: Add files for Simon game challenge
This inlcudes all HTML, CSS and JS code.
The Simon game works by generating a sequence of colors
(or tones) that the player must remember and repeat. Each
round adds a new random color to the sequence. The player
must repeat the full sequence correctly to proceed to the next
round. If the player inputs an incorrect sequence, the game ends and
the player has the option to restart.

Signed-off-by: Joshua Van Straaten <joshuavanstraaten100@gmail.com>